style(hootenanny): fix clippy warnings
- Remove unnecessary `mut` on artifact store references
- Use `is_none_or` instead of `map_or(true, ...)`
- Remove redundant closures in map_err and map
- Fix let_and_return in main.rs
- Remove empty line after doc comment
- Remove unnecessary `as usize` cast

+++ crates/hootenanny/src/mcp_tools/rustysynth.rs
@@ -236,7 +236,7 @@ pub fn inspect_soundfont(soundfont_bytes: &[u8], include_drum_map: bool) -> Resu
                 for region in preset.get_regions() {
                     let key_lo = region.get_key_range_start();
                     let key_hi = region.get_key_range_end();
-                    let instrument_idx = region.get_instrument_id() as usize;
+                    let instrument_idx = region.get_instrument_id();
 
                     let instrument_name = if instrument_idx < instruments.len() {
                         instruments[instrument_idx].get_name().to_string()
